test: fix creation of "std::string"s with \0s
A string literal `"abc"` contains a terminating `\0`, so that is 4
bytes. There is no need to write `"abc\0"` unless two terminating
`\0`s are necessary.

`std::string` objects do not internally contain a terminating `\0`, so
`std::string("abc")` creates a string with size 3 and is the same as
`std::string("abc", 3)`.

In `"\01"` the `01` part is interpreted as one number and that is a
string like `{1, 0}` whereas `"\0z"` is a string like `{0, 'z', 0}`.
To create a string like `{0, '1', 0}` one must use `"\0" "1"`.

Adjust the tests accordingly.